wip on integration of the mosquitto lib + yaml-cpp lib

+++ comau_c5gopen_lpc/include/comau_c5gopen_lpc/c5gopen_driver.h
@@ -142,7 +142,7 @@ namespace c5gopen
     void com_thread( );
     void loop_console_thread( );
     int  initialize_control_position( void );
-    void set_exit_from_open( const int& iArm );
+    void set_exit_from_open( const int8_t& iArm );
 
   };
